SimpCfg: Move testing code into its own file in tests
Also set functions to inline or static as appropriate
