+++ tests/CMakeLists.txt
@@ -118,6 +118,7 @@ llama_target_and_test(test-sampling.cpp)
 llama_target_and_test(test-chat-template.cpp)
 llama_target_and_test(test-chat-template-chaton.cpp)
 llama_target_and_test(test-chaton-groupkv.cpp)
+llama_target_and_test(test-chaton-simpcfg.cpp)
 
 llama_target_and_test(test-grammar-parser.cpp)
 llama_target_and_test(test-llama-grammar.cpp)
